Add instrument example
The instrument has multiple parts:
Instrument hardware - simulated here by a file
Driver - communicates with the instrument
Acquirer - acquires data from the instrument (via the driver) and publishes
the data for other components to use
Controller - Logic for controlling the instrument (e.g. set temperature).
The Controller uses the data from the Acquirer.
Recorder - Records data into a file. The data comes from the Acquirer
DataStore - Temporary storage of data for use by other components
Logger - A logging component that logs data at a slower rate than the
Acquirer produces. This would be implemented in the main program.
Front End - Graphical part of the instrument that shows data and controls
the rest of the components
Status - A GUI window to display the status of each component and turn them on/off

The whole system is fairly robust to components dying and being restarted.